// server/types.ts
export interface Comment {
  _id?: ObjectId;
  text: string;
  commentBy: string;
  commentDateTime: Date;
}

/**
 * Interface extending the request body when adding a comment to a question or an answer, which contains:
 * - id - The unique identifier of the question or answer being commented on.
 * - type - The type of the comment, either 'question' or 'answer'.
 * - comment - The comment being added.
 */
export interface AddCommentRequest extends Request {
  body: {
    id: string; // Unique identifier of the question or answer
    type: 'question' | 'answer'; // Type of the comment target
    comment: {
      text: string; // Content of the comment
      commentBy: string; // Username of the commenter
      commentDateTime: Date; // Date and time of the comment
    };
  };
}